@@ -54,7 +54,7 @@
     "@typescript-eslint/eslint-plugin": "5.5.0",
     "@typescript-eslint/parser": "5.5.0",
     "cross-env": "^7.0.3",
-    "eslint": "8.3.0",
+    "eslint": "8.4.0",
     "husky": "^7.0.4",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
