@@ -17,7 +17,7 @@ export const SCALE_RENDER_STATION_ELEMENTS = DEFAULT_SCALE * Math.pow(ZOOM_VELOC
 export const SCALE_REDUCED_RENDER = SCALE_RENDER_STATION_ELEMENTS / Math.pow(ZOOM_VELOCITY, 40);
 
 // Pan
-export const MAX_PAN_VELOCITY = 30;
+export const MAX_PAN_VELOCITY = 25;
 export const PAN_TRIGGER_LIMIT = 10;
 export const PAN_DECAY_RATE = .9;
 
